+++ packages/svelte/scripts/generate-dts.js
@@ -1,4 +1,18 @@
+import * as fs from 'fs';
 import { createBundle } from 'dts-buddy';
+
+// It may look weird, but the imports MUST be ending with index.js to be properly resolved in all TS modes
+for (const name of ['action', 'animate', 'easing', 'motion', 'store', 'transition']) {
+	fs.writeFileSync(`${name}.d.ts`, `import './types/index.js';`);
+}
+
+fs.writeFileSync('index.d.ts', `import './types/index.js';`);
+fs.writeFileSync('compiler.d.ts', `import './types/index.js';`);
+
+// TODO: some way to mark these as deprecated
+fs.mkdirSync('./types/compiler', { recursive: true });
+fs.writeFileSync('./types/compiler/preprocess.d.ts', `import '../index.js';`);
+fs.writeFileSync('./types/compiler/interfaces.d.ts', `import '../index.js';`);
 
 await createBundle({
 	output: 'types/index.d.ts',
@@ -6,6 +20,7 @@ await createBundle({
 		svelte: 'src/runtime/public.d.ts',
 		'svelte/compiler': 'src/compiler/public.d.ts',
 		'svelte/types/compiler/preprocess': 'src/compiler/preprocess/public.d.ts',
+		'svelte/types/compiler/interfaces': 'src/compiler/interfaces.d.ts',
 		'svelte/action': 'src/runtime/action/public.d.ts',
 		'svelte/animate': 'src/runtime/animate/public.d.ts',
 		'svelte/easing': 'src/runtime/easing/index.js',
@@ -14,3 +29,11 @@ await createBundle({
 		'svelte/transition': 'src/runtime/transition/public.d.ts'
 	}
 });
+
+// There's no way to tell in JS that a class can have arbitrary properties, so we need to add that manually
+const types = fs.readFileSync('types/index.d.ts', 'utf-8');
+fs.writeFileSync(
+	'types/index.d.ts',
+	// same line to not affect source map
+	types.replace(/export class SvelteComponent<[^{]*{/, '$& [prop: string]: any;')
+);