+++ sites/svelte.dev/src/routes/_components/Demo.svelte
@@ -31,8 +31,12 @@
 	let selected = examples[0];
 </script>
 
+<svelte:head>
+	<link rel="prefetch" href="/fonts/overpass/overpass-latin-400.woff2" />
+</svelte:head>
+
 <Section --background="var(--sk-back-2)">
-	<h3>Why Svelte?</h3>
+	<h3>build with ease</h3>
 
 	<div class="container">
 		<div class="controls">